requirements.txt
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
Click==7.1.2
cryptography==3.1
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psutil==5.7.2
pycparser==2.20
PyMySQL==0.10.0
pymysql-pooling==1.0.6
requests==2.24.0
six==1.15.0
Unidecode==1.1.1
urllib3==1.25.10
Werkzeug==1.0.1
redis==3.5.3
XlsxWriter==1.3.3
rq==1.5.1
python-dotenv==0.14.0
xlrd==1.2.0
coloredlogs==14.0
PyJWT==1.7.1
Flask-Mail==0.9.1
blinker==1.4
Markdown==3.2.2
